Start improving the way trackees are added
Clicking or boxing something in the TrackPanel while no trackee is
selected will cause a new trackee to be added using the string from the
text control now. The smallest positive integer (not zero) not used to
identidy a trackee will also be inserted into the text control after
adding a trackee.