Fix find command used to locate dependency files
It implicitly searched the directory tree rooted at the current
directory.  This broke since the directory of the dependency files is
no longer a subdirectory of the makefile's directory.  The format string
used to print the file names is no longer necessary.

Note: the find command used was already flawed before moving the
makefile because it would find .d files in build/release as well as in
build/debug etc.  This wasn't an issue, though, because its results are
filtered to match existing .cpp files.